When building node list from side list, first clear node containers
To me it's all in the name...we're building a node list from the
side list, and hence the former should depend only on the latter.
Consequently the node lists should be cleared before looping
over the sides. I had to do this for some GeoemetricSearch work
in MOOSE combined with mesh coarsening. The node list was containing
some nodes that did not exist on the coarsened mesh and it was
because the BoundaryInfo node containers were not getting cleared